pyproject.toml
[tool.poetry]
name = "dns-mcp-server"
version = "0.1.0"
description = "DNS OSINT MCP Server - Comprehensive DNS reconnaissance tools for threat intelligence"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "dns_mcp_server"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastmcp = "^0.2.0"
dnspython = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"

[tool.poetry.scripts]
dns-mcp-server = "dns_mcp_server.server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88
